use plug._plain_dict() before returning from analyse()
We're experiencing significant delays when we have large objects as plugin instance variables. This edit seems to fix the problem.

It looks like the plugin instance is passed (inside the response object) as-is to flask: perhaps flask is walking through it or duplicating it in memory or some such? Not sure...

There may be a better way?